% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{ambaum_equation}
\alias{ambaum_equation}
\title{Calculates the water vapor saturation pressure as a function of temperature.}
\usage{
ambaum_equation(temp)
}
\arguments{
\item{temp}{temperature in Kelvin (K).}
}
\value{
a vector of the same length as temp with the calculated
water vapor saturation pressure in Pa for each temperature.
}
\description{
Calculates the water vapor saturation pressure as a function of temperature.
}
