% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{ambaum_slope}
\alias{ambaum_slope}
\title{Calculates the slope of the equation of Tetens}
\usage{
ambaum_slope(temp)
}
\arguments{
\item{temp}{temperature in Kelvin (K).}
}
\value{
a vector with the same length as temp with the calculated
slope in Pa/K.
}
\description{
Calculates the slope of the equation of Tetens
}
