% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{calc_atmp}
\alias{calc_atmp}
\title{Calcula a pressão atmosférica com base apenas na altitude}
\usage{
calc_atmp(z)
}
\arguments{
\item{z}{altitude do local, em metros acima do nível do mar}
}
\description{
Calcula a pressão atmosférica com base apenas na altitude
}
