% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{day_length}
\alias{day_length}
\title{Calcula a duração do dia}
\usage{
day_length(j, lat)
}
\arguments{
\item{j}{dia do ano (inteiro de 1 a 366), Date ou POSIXct}

\item{lat}{latitude, em graus decimais (valores negativos para hemisfério sul)}
}
\value{
duração do dia, em horas decimais.
}
\description{
Calcula a duração do dia para a data informada e a latitude
}
