% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{ext_rad_day}
\alias{ext_rad_day}
\title{Calcula a radiação solar exoatmosférica}
\usage{
ext_rad_day(j, lat)
}
\arguments{
\item{j}{dia do ano (de 1 a 366) ou Date ou POSIXct.}

\item{lat}{latitude, em graus decimais (valores negativos para o hemisfério sul)}
}
\description{
Calcula a radiação solar extraterrestre para o dia e a latitude informadas
}
