% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{fao_et0}
\alias{fao_et0}
\title{Calcula a evapotranspiração de referência pelo método de Penman padronizado pela FAO com ou sem as correções propostas pela
American Society of Civil Engineers (ASCE)}
\usage{
fao_et0(
  temp,
  rn,
  g,
  gamma,
  u2,
  es,
  e,
  hourly = FALSE,
  mode = "ASCE-ET0",
  day = NA
)
}
\arguments{
\item{temp}{temperatura média do ar (ºC) (ver detalhes abaixo)}

\item{rn}{saldo de radiação [MJ/m²]}

\item{g}{fluxo de calor no solo [MJ/m²] (ver detalhes)}

\item{gamma}{coeficiente psicrométrico (ver detalhes)}

\item{es}{pressão de saturação do vapor de água (kPa) (ver detalhes)}

\item{e}{pressão parcial do vapor de água (kPa) (ver detalhes)}

\item{hourly}{FALSE para calcular a evapotranspiração diária (mm/dia) (valor padrão), ou TRUE para
calcular a evapotranspiração horária (mm/h)}
}
\description{
Calcula a evapotranspiração de referência (ET0, mm) pelo método de Penman-Monteith
 padronizado pela FAO (Allen et al., 1998) com ou sem as alterações propostas pela American Society of Civil Engineers (ASCE, 2005)
}
\details{
Para fins de padronização, no cálculo para intervalos de tempo de 1 dia ou maiores, deve-se definir temp
 como a média entre a temperatura máxima
 e a temperatura mínima do ar medidas no período de tempo em que se deseja calcular a evapotranspiração,
 e não como a média de todas as medições de temperatura no período (Allen et al., 1998).
 Caso não se disponha de medições diretas do parâmetro rn, pode-se estimar o seu
 valor por meio das funções [net_shortwave] e [net_longwave].
 Caso não se disponha de medições diretas do parâmetro g, Allen et al. (1998) recomendam
 atribuir o valor de 0 para cálculos de evapotranspiração diária ou 0.1 * rn durante o dia
 e 0.5 * rn durante a noite, para cálculos de evapotranspiração horária.
 O valor de es pode ser calculado por meio da função tetens_equation. Recomenda-se
 definir es como a média entre os valores retornados por tetens_equation para
 a temperatura mínima e para a temperatura máxima no período, para cálculos realizados
 no intervalo de tempo de 1 dia ou maior.
}
\references{
ALLEN, R. G.; PEREIRA, L. S.; RAES, D.; SMITH, M. FAO Irrigation and Drainage Paper no 56. FAO: Rome, 1998.
ALLEN, et al.. A recommendation on standardized surface resistance for hourly calculation of reference ETo
 by the FAO56 Penman-Monteith method. Agricultural Water Management, 81, p. 1-22, 2006.
 Available from: <https://www.sciencedirect.com/science/article/abs/pii/S037837740500154X>
}
