% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{inv_rel_dist_sun}
\alias{inv_rel_dist_sun}
\title{Calcula a distância inversa relativa entre a Terra e Sol}
\usage{
inv_rel_dist_sun(j)
}
\arguments{
\item{j}{numeric representando dia do ano, entre 1 e 365, Date ou POSIXt.}
}
\description{
A distância relativa inversa entre a Terra e o Sol é um número
variando de 0,967 UA no afélio a 1,033 UA no periélio.
}
\details{
A Terra orbita o Sol em uma trajetória aproximadamente elíptica. O momento em
que a Terra está mais próxima do Sol é chamado periélio, e o momento onde a distância
é máxima é chamado de afélio. Quanto mais próximo do limite superior (1,033),
maior a proximidade entre a Terra e o Sol.
}
