% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{landsat8_read_meta}
\alias{landsat8_read_meta}
\title{Reads landsat metadata from a json file and converts it to a more accessible
format that R can understand.}
\usage{
landsat8_read_meta(filename)
}
\description{
A short description...
}
