% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{log_wind_profile}
\alias{log_wind_profile}
\title{Estimates wind speed at height z2 based on wind speed measued at
height z1, given the height of the zero plane displacement and the roughness
length of the surface.}
\usage{
log_wind_profile(u1, z2 = 2, z1 = 10, d = 0.084, z0 = 0.01476)
}
\arguments{
\item{u1}{Wind speed [m/s] measured at height z1.}

\item{z2}{Height z2 [m].}

\item{z1}{Height z1 [m]. The default is 10.}

\item{d}{Zero plane displacement height. If not set, a default value = 0.08 will be set, which
is the zero plane displacement height estimated for a 0.12m height uniform crop.}

\item{z0}{Roughness length. If not set, a default value of 0.01476 will be set, which
corresponds to the roughness length of the standardized FAO ET0 Penman-Monteith
equation.}
}
\value{
Wind speed at height z2.
}
\description{
For a standardized FAO Peman Monteith ET0 surface
you can use the default values for d and z0. If the wind speed is measured
at a standard weather station, which measures wind speed
at a 10m height, you can use the default value for z1.
}
