% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric}
\alias{metric}
\title{Creates a object that represents a model with its data.}
\usage{
metric(rast, terrain, dt, weather_station = NULL)
}
\arguments{
\item{rast}{raster data. Must contain, at least, NDVI, SAVI, albedo and surface temperature.}
}
\description{
Creates a object that represents a model with its data.
}
