% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_albedo_l8}
\alias{metric_albedo_l8}
\title{Calculates the albedo from a landsat 8 image using the procedure described by
Silva et al. (2016).}
\usage{
metric_albedo_l8(b2, b3, b4, b5, b6, b7, tau, alpha_atm = 0.03)
}
\arguments{
\item{b2}{Landsat 8 top-of-atmosphere reflectance at band 2}

\item{b3}{Landsat 8 top-of-atmosphere reflectance at band 3}

\item{b4}{Landsat 8 top-of-atmosphere reflectance at band 4}

\item{b5}{Landsat 8 top-of-atmosphere reflectance at band 5}

\item{b6}{Landsat 8 top-of-atmosphere reflectance at band 6}

\item{b7}{Landsat 8 top-of-atmosphere reflectance at band 7}

\item{tau}{atmospheric transmissivity}

\item{alpha_atm}{atmospheric albedo (see details)}
}
\description{
Calculates the broadband albedo through the integration of narrow band reflectance
for bands 2, 3, 4, 5, 6, and 7 of Landsat 8.
}
\details{
The atmospheric albedo can be obtained from a radiative transfer model, such as
MODTRAN. It is usually a value between 0.025 and 0.040. Here we use a 'default'
value of 0.03, but we recommend you to adapt the value according to your needs.
}
