% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_get_coordinates}
\alias{metric_get_coordinates}
\title{Retrieves the longitude and latitude of each raster cell in a new raster, regardless
of the coordinate reference system of the source image.}
\usage{
metric_get_coordinates(src)
}
\arguments{
\item{src}{source image}
}
\description{
Retrieves the longitude and latitude of each raster cell in a new raster, regardless
of the coordinate reference system of the source image.
}
