% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_horizontal_incidence_angle}
\alias{metric_horizontal_incidence_angle}
\title{Calculates the solar incidence angle over a horizontal surface.}
\usage{
metric_horizontal_incidence_angle(lat, lon, datetime)
}
\description{
Calculates the solar incidence angle over a horizontal surface.
}
