% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_hour_angle}
\alias{metric_hour_angle}
\title{Calculates the hour angle at a given time based on the longitude.}
\usage{
metric_hour_angle(lon, dt)
}
\arguments{
\item{lon}{SpatRaster containing the longitude of each raster cell.}
}
\description{
Calculates the hour angle at a given time based on the longitude.
}
