% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_leaf_area_index}
\alias{metric_leaf_area_index}
\title{Calculates the leaf area index based on SAVI.}
\usage{
metric_leaf_area_index(x)
}
\arguments{
\item{savi}{soil adjusted vegetation index.}
}
\description{
Calculates the leaf area index based on SAVI.
}
