% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_ndvi}
\alias{metric_ndvi}
\title{Metric NDVI}
\usage{
metric_ndvi(red, nir)
}
\arguments{
\item{red}{reflectance on the red band.}

\item{nir}{reflectance on the near-infrared band.}
}
\description{
Calculates the Normalized Difference Vegetation Index.
}
\details{
The normalizeed difference vegetation Index is calculated as
NDVI = (nir - red) / (nir + red).
Which results in a number between -1 and 1.
}
