% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_relative_distance_earth_sun}
\alias{metric_relative_distance_earth_sun}
\title{Calculates the square of the relative distance Earth-Sun.}
\usage{
metric_relative_distance_earth_sun(dt)
}
\description{
Calculates the square of the relative distance Earth-Sun.
}
