% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric_savi}
\alias{metric_savi}
\title{Calculates the soil adjusted vegetation index (SAVI).}
\usage{
metric_savi(red, nir, const = 0.1)
}
\arguments{
\item{red}{top-of-atmosphere reflectance on the red band.}

\item{nir}{top-of-atmosphere reflectance on the near-infrared band.}

\item{const}{numeric constant usually set to 0.1 for METRIC applications.}
}
\description{
Calculates the soil adjusted vegetation index (SAVI).
}
