% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{net_longwave}
\alias{net_longwave}
\title{Calcula o saldo de radiação de ondas longas}
\usage{
net_longwave(tmax, tmin, e, rs, rs0, hourly = FALSE)
}
\arguments{
\item{tmax}{Temperatura máxima do ar medida dentro do intervalo de tempo do cálculo [ºC]}

\item{tmin}{Temperatura mínima do ar medida dentro do intervalo de tempo do cálculo [ºC]}

\item{e}{Pressão parcial do vapor [kPa] média dentro do intervalo de tempo do cálculo.}

\item{rs}{Radiação solar de onda curta incidente sobre uma superfície horizontal, durante o intervalo de tempo do cálculo [MJ/m²]}

\item{rs0}{Radiação solar de onda curta em condições de céu claro [MJ/m²]}

\item{hourly}{TRUE para intervalo de tempo de 1 hora, FALSE caso contrário.}
}
\description{
Calcula a radiação de ondas longas conforme metodologia
proposta por Allen et al. (1998)
}
\references{
ALLEN, R. G.; PEREIRA, L. S.; RAES, D.; SMITH, M. FAO Irrigation and Drainage Paper no 56. FAO: Rome, 1998.
}
