% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{net_shortwave}
\alias{net_shortwave}
\title{Calculates net shortwave radiation}
\usage{
net_shortwave(rs, albedo = 0.23)
}
\arguments{
\item{rs}{global radiation (W/m², MJ/m²/d, kJ/m², etc...)}

\item{albedo}{surface albedo (adimensional) between 0 and 1}
}
\description{
Calculates net shortwave radiation, which is the difference between
incoming shortwave radiation and reflected shortwave radiation over a surface.
}
\details{
O parâmetro albedo possui valor padrão de 0.23. Este é o albedo para a cultura
 de referência, utilizado no cálculo da evapotranspiração de referência. Para maiores detalhes
 veja a seção 'details' da função \code{\link{fao_et0}}.
}
