% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{psy_const}
\alias{psy_const}
\title{Calculates the psychrometric constant}
\usage{
psy_const(patm)
}
\arguments{
\item{patm}{atmospheric pressure (Pa).}
}
\value{
Psychrometric constant (Pa / K)
}
\description{
Calculates the psychrometric constant
}
