% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{solar_declination}
\alias{solar_declination}
\title{Calcula a declinação solar}
\usage{
solar_declination(j)
}
\arguments{
\item{j}{dia do ano (1 a 365 ou 366), objeto Date ou POSIXt}
}
\description{
Calcula a declinação solar. Declinação solar (δ) é o ângulo
formado entre o Sol e o plano do equador terrestre.
}
\details{
Ao longo do ano, a
declinação solar varia de -23.44° a 23.44° (em radianos: -0.4091052 rad a
+0.4091052 rad), com valores negativos ocorrendo no verão do hemisfério sul e
valores positivos no verão do hemisfério norte. A declinação solar é igual a
0 durante os equinócios.
}
