% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{stefan_boltzmann_law}
\alias{stefan_boltzmann_law}
\title{Calculates the radiant emittance of a body as a function of its temperature}
\usage{
stefan_boltzmann_law(temp, epsilon = 1)
}
\arguments{
\item{temp}{absolute temperature (K)}

\item{epsilon}{body's emissivity (adimensional between 0 and 1). A value of 1
calculates black body radiant emittance.}
}
\description{
Calculates the radiant emittance of a body as a function of its
 absolute temperature.
}
