% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{stress_logistic}
\alias{stress_logistic}
\title{Fits a logistic function to describe a stress response}
\usage{
stress_logistic(sn, sx)
}
\description{
Fits a logistic function to describe a stress response
}
