% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{stress_spline}
\alias{stress_spline}
\title{Stress response based on a spline function}
\usage{
stress_spline(known_srel, known_ks, method = "hyman", ties = mean)
}
\description{
Stress response based on a spline function
}
