% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{tetens_equation}
\alias{tetens_equation}
\title{Calculates the water vapor saturation pressure as a function of temperature.}
\usage{
tetens_equation(temp)
}
\arguments{
\item{temp}{temperature in Celsius (ºC). Rasters are also supported.}
}
\value{
a vector of the same length as temp with the calculated
water vapor saturation pressure in kPa for each temperature.
}
\description{
Calculates the water vapor saturation pressure as a function of temperature.
}
\examples{
> tetens_equation(seq(5, 30, by = 5))
[1] 0.872311 1.227963 1.705346 2.338281 3.167778 4.243065
}
