% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{tetens_slope}
\alias{tetens_slope}
\title{Calculates the slope of the equation of Tetens}
\usage{
tetens_slope(temp)
}
\arguments{
\item{temp}{temperature in Celsius (ºC). Rasters are also supported.}
}
\value{
a vector of the same length as temp with the calculated
slope in kPa/ºC.
}
\description{
Calculates the slope of the equation of Tetens
}
