% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite.R
\name{thornthwaite_pet}
\alias{thornthwaite_pet}
\title{Calcula a evapotranspiração potencial (ETP) por meio do método de Thornthwaite e Mather}
\usage{
thornthwaite_pet(temp, ...)
}
\arguments{
\item{temp}{a vector containing 12 values for air temperature or a raster::RasterStack
with 12 layers of air temperature.}

\item{...}{other arguments}

\item{lat}{latitude}
}
\description{
Calcula a evapotranspiração potencial (ETP) por meio do método de Thornthwaite e Mather
}
