% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite.R
\name{water_balance}
\alias{water_balance}
\title{Calculates the water balance using the method of Thornthwaite & Mather}
\usage{
water_balance(
  p,
  pet,
  taw,
  ...,
  aw_initial = taw,
  max_iter = 256,
  max_error = 0,
  sequential = FALSE
)
}
\arguments{
\item{p}{precipitation}
}
\description{
Calculates the water balance using the method of Thornthwaite & Mather
}
